requirements.txt
requests
beautifulsoup4
pywebview
